fix: change architecture
